refactor(planner): move getSquareCompanionStatus to separate utility
Move helper function to src/utils/companionStatus.js to fix
react-refresh/only-export-components ESLint error. Component files
should only export components for fast refresh to work properly.
