+++ square-gardener/src/components/PlanningGrid.jsx
@@ -1,7 +1,8 @@
 import { useMemo, useState } from 'react';
 import PropTypes from 'prop-types';
 import { getPlantById } from '../data/plantLibrary';
-import { validateArrangement, getAdjacentPositions } from '../utils/planningAlgorithm';
+import { validateArrangement } from '../utils/planningAlgorithm';
+import { getSquareCompanionStatus } from '../utils/companionStatus';
 
 const PLANT_COLORS = {
   tomato: '#ef4444',
@@ -46,58 +47,6 @@ const getContrastColor = (hexColor) => {
   const b = parseInt(hexColor.slice(5, 7), 16);
   const luminance = (0.299 * r + 0.587 * g + 0.114 * b) / 255;
   return luminance > 0.5 ? '#1f2937' : '#ffffff';
-};
-
-/**
- * Get companion/enemy status for a square
- * @param {Array<Array<string|null>>} grid - The garden grid
- * @param {number} row - Row index
- * @param {number} col - Column index
- * @returns {{ hasCompanion: boolean, hasEnemy: boolean, companions: string[], enemies: string[] }}
- */
-export const getSquareCompanionStatus = (grid, row, col) => {
-  const plantId = grid[row]?.[col];
-  if (!plantId) {
-    return { hasCompanion: false, hasEnemy: false, companions: [], enemies: [] };
-  }
-
-  const plant = getPlantById(plantId);
-  if (!plant) {
-    return { hasCompanion: false, hasEnemy: false, companions: [], enemies: [] };
-  }
-
-  const height = grid.length;
-  const width = grid[0].length;
-  const adjacentPositions = getAdjacentPositions(row, col, width, height);
-
-  const companions = [];
-  const enemies = [];
-
-  for (const pos of adjacentPositions) {
-    const adjacentPlantId = grid[pos.row][pos.col];
-    if (!adjacentPlantId) continue;
-
-    const adjacentPlant = getPlantById(adjacentPlantId);
-    if (!adjacentPlant) continue;
-
-    if (plant.companionPlants.includes(adjacentPlantId)) {
-      if (!companions.includes(adjacentPlant.name)) {
-        companions.push(adjacentPlant.name);
-      }
-    }
-    if (plant.avoidPlants.includes(adjacentPlantId)) {
-      if (!enemies.includes(adjacentPlant.name)) {
-        enemies.push(adjacentPlant.name);
-      }
-    }
-  }
-
-  return {
-    hasCompanion: companions.length > 0,
-    hasEnemy: enemies.length > 0,
-    companions,
-    enemies
-  };
 };
 
 function PlanningGrid({ arrangement, bed, onSquareClick, lockedSquares, onArrangementChange, editable = false }) {
